airflow_app/pyproject.toml
[tool.poetry]
name = "airflow-etl"
version = "0.1.0"
description = ""
authors = ["Sean Deery <sdeery14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
apache-airflow = "^2.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
